requirements.txt
appdirs==1.4.4
awscli==1.18.213
beautifulsoup4==4.9.3
black==20.8b1
boto3==1.16.53
botocore==1.19.53
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.3
docutils==0.15.2
idna==2.10
jmespath==0.10.0
mypy-extensions==0.4.3
oandapyV20==0.6.3
pathspec==0.8.1
pyasn1==0.4.8
python-dateutil==2.8.1
python-dotenv==0.15.0
PyYAML==5.3.1
regex==2020.11.13
requests==2.25.1
rsa==4.5
s3transfer==0.3.4
six==1.15.0
soupsieve==2.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.2
